+++ CHANGELOG.md
@@ -8,6 +8,7 @@ UNRELEASED
 * [ [#2065](https://github.com/digitalfabrik/integreat-cms/issues/2065) ] Only allow staff members to use SUMM.AI bulk action
 * [ [#1970](https://github.com/digitalfabrik/integreat-cms/issues/1970) ] Disable drag & drop in page order for users without editing permission
 * [ [#2112](https://github.com/digitalfabrik/integreat-cms/issues/2112) ] Hide form buttons and bulk actions for users without editing permissions
+* [ [#2075](https://github.com/digitalfabrik/integreat-cms/issues/2075) ] Add ability to ignore calculation of HIX value for certain pages
 
 
 2023.2.2